backend/requirements.txt
fastapi
uvicorn
torch
transformers
pdf2image
pillow
numpy
